CRUD na API de coleta de combustível

+++ src/main/java/com/example/selecaojava/repository/CountyRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CountyRepository extends PagingAndSortingRepository<County, Long> {
+public interface CountyRepository extends PagingAndSortingRepository<County, Integer> {
 }
